+++ RELEASE_NOTES.md
@@ -6,6 +6,7 @@ _Date TBD_
 ## Minor Features
 * Improve capturing of database errors on listing upload or update.
 * Improve error handling for invalid length/format codes in the CHPL unique product id.
+* Send a cache-cleared header when the search cache has been evicted
 
 ---